refactor: rename aggregation fns
