+++ src/timeseriesflattener/aggregation_fns.py
