+++ src/timeseriesflattener/tests/test_timeseriesflattener/test_feature_spec_objects.py
@@ -5,7 +5,7 @@ from typing import List
 
 import numpy as np
 
-from timeseriesflattener.aggregation_functions import maximum
+from timeseriesflattener.aggregation_fns import maximum
 from timeseriesflattener.feature_specs.group_specs import (
     NamedDataframe,
     PredictorGroupSpec,
